add basic theme
